requirements.txt
fastapi
uvicorn
blake3
PyNaCl
pydantic
sqlalchemy
psycopg2-binary
pytest
